refactoring: add PartitioningScheme
